Add product CRUD

+++ src/main/java/com/ant_storage/ANT/Storage/service/ProductService.java
@@ -2,6 +2,7 @@ package com.ant_storage.ANT.Storage.service;
 
 import com.ant_storage.ANT.Storage.entity.Product;
 import com.ant_storage.ANT.Storage.repository.ProductRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -25,7 +26,24 @@ public class ProductService {
         return productRepository.save(product);
     }
 
+    public Product updateProduct(Integer id, Product updatedProduct) {
+        Product existingProduct = productRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Product not found with id: " + id));
+
+        existingProduct.setName(updatedProduct.getName());
+        existingProduct.setDescription(updatedProduct.getDescription());
+        existingProduct.setLocation(updatedProduct.getLocation());
+        existingProduct.setDate(updatedProduct.getDate());
+        existingProduct.setCategory_id(updatedProduct.getCategory_id());
+
+        // Save the updated user
+        return productRepository.save(existingProduct);
+    }
+
     public void deleteProduct(Integer id) {
+        if (id == null) {
+            throw new IllegalArgumentException("ID must not be null");
+        }
         productRepository.deleteById(id);
     }
 }
